Separate implementation of URL replacement service into its own file
… that does not have to be loaded into each extension.

Fixes another one of #1055
