+++ src/amp-core-service.js
@@ -19,6 +19,7 @@ import {installFramerateService} from './service/framerate-impl';
 import {installHistoryService} from './service/history-impl';
 import {installResourcesService} from './service/resources-impl';
 import {installStandardActions} from './service/standard-actions-impl';
+import {installUrlReplacementsService} from './service/url-replacements-impl';
 import {installViewerService} from './service/viewer-impl';
 import {installViewportService} from './service/viewport-impl';
 import {installVsyncService} from './service/vsync-impl';
@@ -38,4 +39,5 @@ export function installCoreServices(window) {
   installResourcesService(window);
   installStandardActions(window);
   installFramerateService(window);
+  installUrlReplacementsService(window);
 }
